requirements.txt
autopep8==1.6.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
unicode==2.8
Unidecode==1.3.3
urllib3==1.26.8
